+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+    - [cmd/list] added `--filter` option (GitHub Issue #6)
     - [cmd] added `link` command to change relationships between issues (GitHub Issue #29)
     - [cli] accept `due-date` and `due_date` for `list` columns, `new` and `set` commands
     - [test] added more checks for `basic_workflow` and more tests for `set` command