+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+### Added
+
+    - [settings] added `settings.yaml` to hold local non-version-controlled user settings (GitHub Issue #65)
+
+### Changed
+
+    - [config] moved `editor` from `config.yaml` to `settings.yaml` (GitHub Issue #65)
+
 ## [v0.4.0] - 2025-12-26
 
 ### Added