+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+    - [cmd/set] added shortcut 'me' for `--reporter` and `--assignee`, automatically takes the value from `settings.yaml:user` (GitHub Issue #68)
+    - [cmd/new] added shortcut 'me' for `--reporter` and `--assignee`, automatically takes the value from `settings.yaml:user` (GitHub Issue #68)
     - [meta] added `reporter` field (GitHub Issue #64)
     - [settings] added `settings.yaml` to hold local non-version-controlled user settings (GitHub Issue #65)