+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+    - [config] added possibility to change ID generation strategy `id_generation` (GitHub Issue #9)
+
 ## [v0.5.0] - 2025-12-27
 
 ### Added