+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Added
+
+    - [cargo] MSRV v1.88
+    - [test]  added automated tests for `set` bulk operations
+
 ### Changed
 
     - [cmd/list] harmonized sorting of empty values: First for asc, last for desc (GitHub Issue #95)